requirements.txt
# Core Telegram Bot Dependencies
python-telegram-bot[job-queue]==22.3
python-dotenv==1.1.1
aiofiles==24.1.0
apscheduler==3.11.0

# Browser Automation
patchright==1.52.5
camoufox[geoip]
playwright
DrissionPage==4.0.5.6

# API Server (for Turnstile solver)
quart
hypercorn
flask[async]

# HTTP and Networking (kept for compatibility)
aiohttp==3.10.11
requests==2.32.3

# Data Processing
lxml==5.4.0

# Utilities
psutil==7.0.0
simple-useragent
loguru
pydantic
proxystr
python-socketio

# Image processing for solver
pyautogui
opencv-python

# Dropbox integration
dropbox